feat: handle change role member in workspace
