+++ frontend/src/components/setting/list-member.tsx
@@ -1,7 +1,9 @@
-import { Pane, PaneProps, Table, Image, majorScale, Combobox, Pagination } from 'evergreen-ui'
-import { Member } from '~/models/user'
+import { Pane, PaneProps, Table, Image, majorScale, Combobox, Pagination, toaster } from 'evergreen-ui'
+import { useParams } from 'react-router'
+import { Member, WorkspaceSetRolePayload } from '~/models/member'
 import { PaginationResponse } from '~/models/utils'
-import { WorkspaceRole } from '~/models/workspace'
+import { WorkspaceParams, WorkspaceRole } from '~/models/workspace'
+import workspaceService from '~/services/workspace.service'
 
 interface ListMemberCompProps extends PaneProps {
   members: PaginationResponse<Member>
@@ -12,9 +14,20 @@ interface ListMemberCompProps extends PaneProps {
 
 export const ListMemberComp = (props: ListMemberCompProps) => {
   const { members, handleChangePage, filterByUsername, filterByRole, ...paneProps } = props
-
+  const params = useParams()
   const handleChangeRole = (role: WorkspaceRole, user_id: number) => {
-    console.log(role, user_id)
+    const _params: WorkspaceParams = { permalink: params.permalink || '' }
+    const payload: WorkspaceSetRolePayload = {
+      user_id: user_id,
+      role: role
+    }
+    workspaceService.changeRoleMember(_params, payload).then((data) => {
+      if (data.status === 200) {
+        toaster.success(data.message)
+      } else {
+        toaster.danger(data.message)
+      }
+    })
   }
 
   return (
