+++ frontend/src/models/user.ts
@@ -1,6 +1,3 @@
-import { ProjectMinimize } from './project'
-import { WorkspaceRole } from './workspace'
-
 export interface User {
   id: number
   username: string
@@ -15,14 +12,4 @@ export interface UserDetail extends User {
   phone_number: string | null
   created_at: string
   updated_at: string
-}
-
-export interface Member {
-  avatar: string
-  first_name: string
-  id: number
-  last_name: string
-  role: WorkspaceRole
-  username: string
-  project: ProjectMinimize[]
 }
